Famigo-Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-dotenv==1.0.1
PyJWT==2.9.0
aiosqlite==0.20.0
python-multipart==0.0.9
bcrypt==4.2.0
